Reimplement single-commit authorization method
The previous code granted access to display a commit iif it was an
ancestor of any of the local or remote branches accessible to the
calling request. This exclude authorization use cases where commits
are in refspecs outside the usual namespace practices for commits
in a git repository (i.e. refs/heads/*).

This commit delegates the authorization phase to the new operation
'authorize-commit' to be implemented in the authorization helper
script.